pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dart-mcp-server"
version = "0.1.0"
description = "MCP server for Open DART API integration"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "pandas>=2.2.2", 
    "dart-fss>=0.3.12",
    "python-dotenv>=1.0.1",
    "mcp>=1.0.0"
]

[project.scripts]
dart-mcp-server = "src.dart_mcp_server:main"

[project.urls]
Homepage = "https://github.com/yourusername/dart-mcp-server"
Repository = "https://github.com/yourusername/dart-mcp-server" 